Fixed /robplayer
So I had an issue where targetplayer was coming back nil even though it should've been working. Since it's already been defined on 675 and that triggers when you rob someone I took off the local so that it can be called through the whole resource and defined it up top. Seems to be working so far, if any better way to do it please let me know.