Changes made to using item
Changed use item functionality
- Modified to have a check event prior to progbar triggering. If the item doesnt have or need an event check (burger for example), it just proceeds to trigger progbar with no check.
- Modified to trigger event after progbar is complete
- Modified progbar to allow more customization (labels, ability to cancel).

Example used with Bandage. Event 'bandageCheck' checks if the players heal is not max (200 HP). If it isn't, callback true and proceeds with progbar which has the ability to cancel if accidently used. Once progbar is completed, it triggers to bandage event to heal the player.