feat(server/inventory): provide exports for simpler integration
There are cases where a resource may need to create a stash, but exporting the entire module is excessive.

Adds both CreateStash and ConvertItems as exports.

ConvertItems will take traditional itemdata input and updates it for compatibility with ox_inventory, i.e.

{"cola":1, "bread":3} will convert to [{"slot":1,"name":"cola","count":1}, {"slot":2,"name":"bread","count":3}]

CreateStash is a slightly modified function to call Inventory.Create and will attempt to convert incompatible itemdata if received.