refactor(server/inventory): replace CreateStash with RegisterStash
CreateStash was a poor implementation, since it required entering all the data on the server every time we wanted a stash - which could later get saved and removed from memory.

RegisterStash simply sets up a stash "type" which can be requested by the client. This functionality was already in place but mixed with the old implementation.