refactor: rename player module to bridge
This module is used for more than player data, so set a more appropriate name.
This will be expanded on later to help provide compatibility with different frameworks.

Includes change to player accounts when syncing data.
Any additional accounts should be added to the bridge module to ensure sync between the item and the account.