feat(logs): datadog logging
I'll probably add some other options (such as a LIMITED discord logging setting) in the future. For now, this is the best way to track the often large amount of data generated by ox_inventory.

Will probably refactor later to store logs for a bit before triggering the POST event.