fix(server/inventory): weapon slot syncing
Need to inform client when moving the weapon, as before.
Set inventory.weapon to nil before sending the disarm event to the client.
Check if an item is actually a weapon during the updateWeapon event
and the GetCurrentWeapon export, else set to nil.

Last two cases shouldn't happen, but better safe than sorry.