Changed key value for 'slots' return in search function
searches were only returning one slot for any item searched, no matter how many individual items were in the inventory. Issue was that the key value it was trying to set in the 'returnData[item]' table was always 1. It has been changed to set the slot number as the key