refactor(mysql): new namespace, adjust query formatting
People manage to mess up (alter?) their fxmanifest and not load the
mysql module, then blame me for them getting an error that doesn't
occur when properly using the resource (hooray).

Move the functions to the db "namespace" to make things more obvious.

Also alter the query placeholders and formatting, for readability.