refactor(data/items): remove lockpick usable bits
There's no item callback because people are supposed to add their own, but
ESX.UseItem won't function if the consume or client properties exist on
an item. This didn't matter until the UsableItems table was removed because
certain fuckwits complained about non-reproducible issues, but now they
complain because ESX.UseItem is only called if the consume property
doesn't exist. This change is apparently incomprehensible to some people and "no work".