chore(server): add a comment about errors during server.UseItem
Versions of ESX prior to 1.8.5 do not safely invoke registered item
callbacks, which will throw errors if the function fails to execute due
to an error, the registering resource being stopped, etc.

ESX versions prior to 1.7.5 do not check if the given item was registered
as usable in the first place, trying to call a nil value for any and all items.

There's no way to solve this internally or in a backwards-compatible manner.