feat(client/weapons): add weapon equip anim skip
this adds a new optional param to the useSlot export, it allows for you to disable the anim for weapons if enabled when calling useSlot externally.
while this is a little bit niche, i think this can come in handy in any instances where you need to set a weapon to a player immediately.